requirements.txt
discord
requests
colorama
virustotal-api
hashlib